ApacheCLI

+++ Task3/src/main/java/by/epam/bigdatalab/dao/impl/PostgreSQLDAOImpl.java
@@ -168,7 +168,7 @@ public class PostgreSQLDAOImpl implements DataBaseDAO {
     private List<CrimeOutcomeStatus> getAllCrimeOutcomeStatuses() {
 
         return query.select(GET_OUTCOME_STATUSES)
-                .listResult(resultSet -> new CrimeOutcomeStatus(resultSet.getInt(OUTCOME_STATUS_ID),
+                .listResult(resultSet -> new CrimeOutcomeStatus(resultSet.getLong(OUTCOME_STATUS_ID),
                         resultSet.getString(OUTCOME_STATUS_CATEGORY),
                         resultSet.getDate(OUTCOME_STATUS_DATE)));
     }
@@ -176,7 +176,7 @@ public class PostgreSQLDAOImpl implements DataBaseDAO {
     private List<CrimeLocationStreet> getAllCrimeLocationStreets() {
 
         return query.select(GET_STREETS)
-                .listResult(resultSet -> new CrimeLocationStreet(resultSet.getInt(STREET_ID),
+                .listResult(resultSet -> new CrimeLocationStreet(resultSet.getLong(STREET_ID),
                         resultSet.getString(STREET_NAME)));
     }
 
@@ -185,10 +185,10 @@ public class PostgreSQLDAOImpl implements DataBaseDAO {
 
         return query.select(GET_LOCATIONS)
                 .listResult(resultSet ->
-                        new CrimeLocation(resultSet.getInt(LOCATION_ID),
+                        new CrimeLocation(resultSet.getLong(LOCATION_ID),
                                 resultSet.getDouble(LOCATION_LATITUDE),
                                 resultSet.getDouble(LOCATION_LONGITUDE),
-                                getCrimeLocationStreet(crimeLocationStreets, resultSet.getInt(LOCATION_STREET_ID))));
+                                getCrimeLocationStreet(crimeLocationStreets, resultSet.getLong(LOCATION_STREET_ID))));
     }
 
 
@@ -198,20 +198,20 @@ public class PostgreSQLDAOImpl implements DataBaseDAO {
 
         return query.select(GET_CRIMES)
                 .listResult(resultSet ->
-                        new Crime(resultSet.getInt(CRIMES_ID),
+                        new Crime(resultSet.getLong(CRIMES_ID),
                                 resultSet.getString(CRIMES_CATEGORY),
                                 resultSet.getString(CRIMES_PERSISTENT_ID),
                                 resultSet.getDate(CRIMES_MONTH),
                                 resultSet.getString(CRIMES_CONTEXT),
                                 resultSet.getString(CRIMES_LOCATION_TYPE),
                                 resultSet.getString(CRIMES_LOCATION_SUBTYPE),
-                                getCrimeLocation(crimeLocations, resultSet.getInt(CRIMES_LOCATION_ID)),
-                                getCrimeOutcomeStatus(crimeOutcomeStatuses, resultSet.getInt(CRIMES_OUTCOME_STATUS_ID)
+                                getCrimeLocation(crimeLocations, resultSet.getLong(CRIMES_LOCATION_ID)),
+                                getCrimeOutcomeStatus(crimeOutcomeStatuses, resultSet.getLong(CRIMES_OUTCOME_STATUS_ID)
                                 )));
 
     }
 
-    private CrimeLocationStreet getCrimeLocationStreet(List<CrimeLocationStreet> crimeLocationStreets, int streetId) {
+    private CrimeLocationStreet getCrimeLocationStreet(List<CrimeLocationStreet> crimeLocationStreets, long streetId) {
         for (CrimeLocationStreet street : crimeLocationStreets) {
             if (street.getId() == streetId) {
                 return street;
@@ -220,7 +220,7 @@ public class PostgreSQLDAOImpl implements DataBaseDAO {
         return null;
     }
 
-    private CrimeLocation getCrimeLocation(List<CrimeLocation> crimeLocations, int crimeLocationId) {
+    private CrimeLocation getCrimeLocation(List<CrimeLocation> crimeLocations, long crimeLocationId) {
 
         for (CrimeLocation crimeLocation : crimeLocations) {
             if (crimeLocation.getId() == crimeLocationId) {
@@ -230,7 +230,7 @@ public class PostgreSQLDAOImpl implements DataBaseDAO {
         return null;
     }
 
-    private CrimeOutcomeStatus getCrimeOutcomeStatus(List<CrimeOutcomeStatus> crimeOutcomeStatuses, int crimeOutcomeId) {
+    private CrimeOutcomeStatus getCrimeOutcomeStatus(List<CrimeOutcomeStatus> crimeOutcomeStatuses, long crimeOutcomeId) {
 
         for (CrimeOutcomeStatus crimeOutcomeStatus : crimeOutcomeStatuses) {
             if (crimeOutcomeStatus.getId() == crimeOutcomeId) {
@@ -240,7 +240,7 @@ public class PostgreSQLDAOImpl implements DataBaseDAO {
         return null;
     }
 
-    private int getCrimeLocationId(CrimeLocation crimeLocation) {
+    private long getCrimeLocationId(CrimeLocation crimeLocation) {
         List<CrimeLocation> crimeLocations = new ArrayList<>(getAllCrimeLocations());
         for (CrimeLocation crimeLoc : crimeLocations) {
             if (crimeLoc.equals(crimeLocation)) {
@@ -250,7 +250,7 @@ public class PostgreSQLDAOImpl implements DataBaseDAO {
         return -1;
     }
 
-    private int getCrimeOutcomeStatusId(CrimeOutcomeStatus crimeOutcomeStatus) {
+    private long getCrimeOutcomeStatusId(CrimeOutcomeStatus crimeOutcomeStatus) {
 
         List<CrimeOutcomeStatus> crimeOut = new ArrayList<>(getAllCrimeOutcomeStatuses());
         for (CrimeOutcomeStatus status : crimeOut) {
